Split out the timeline chart code into `TextTimelineChart`.
- All `History` classes provided `toTimelineChart`, using this new class
  and default settings.
- But this would allow customization of the chart by the client.
